backend/Services/requirements.txt
asgiref==3.8.1
Django==5.2.1
django-cors-headers==4.7.0
django-jazzmin==3.0.1
django-tailwind==4.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
openpyxl==3.1.5
PyJWT==2.9.0
sqlparse==0.5.3
